refactor : TestDeckForThreeParticipants 클래스를 테스트 패키지로 이동
