+++ src/test/java/domain/game/BetTest.java
@@ -1,6 +1,5 @@
 package domain.game;
 
-import domain.deck.TestDeckForThreeParticipant;
 import domain.participant.Dealer;
 import domain.participant.Name;
 import domain.participant.Participant;
@@ -13,7 +12,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
